Made it possible to generate more than 255 uuid's

+++ src/main.rs
@@ -6,7 +6,7 @@ use uuid::Uuid;
 struct Args {
     /// Number of uuids to generate
     #[arg(short, long, default_value_t = 1)]
-    number: u8,
+    number: u16,
 }
 
 fn main() {
